requirements.txt
flask==1.0.2
flask-restplus==0.9.2
flask-sqlalchemy==2.3.0
marshmallow==2.18.1
psycopg2==2.7.6.1
numpy
scipy
pandas
scikit_learn
